Allow passing SQL expressions to SET for UPDATEs
This allows us to update a row based on values it contains, avoiding
race conditions involved in loading data into application memory,
performing the computation, and saving the result back to the DB.

For example, consider the following:

    record = query.find_by_id(id)
    query.increment_count for: record

In that example, the record is fetched from the DB with `find_by_id`. In
the second line, we compute a new count and save that value back to the
database. But without row locks (which we aren't guaranteed to be using
here), the count could change in between those two lines of code, in
which case we would end up saving the wrong count. If, instead, we
update the record without retrieving it, this becomes an atomic update
without a row lock:

    query.increment_count(id)