Add experimental support for "fake" queries
"Fake" queries are ones where you're not actually querying a real table,
just an expression. This is useful for querying generated data without
persisting it to a table while still allowing it to be joined with other
expressions.

This is highly experimental for now, so don't rely on it for anything
critical. It may not even stick around.