Improve support for multi-model queries
This commit makes it so you don't have to resort to weird things like
overriding `self.new` to accommodate the JOIN clause. Instead, you can
replace the `table` macro with a `from` and an `inner_join`:

struct ArticleFeedQuery < Interro::QueryBuilder({Article, User})
  from "articles",
    inner_join("users", as: "author", on: "articles.author_id = author.id")
end

I don't know if this DSL will stick around as-is, but it works for now.