Add Any type
This type is similar to JSON::Any and YAML::Any in the stdlib, wrapping
other data types to avoid recursive aliases. The Crystal type checker
really struggles with recursive aliases and Crystal 1.6 broke some of
the ways we use them, so we'll have to deal with the intermediate type
wrapper.